Renames Controller to Config since all classes have been SRPd. Adjusts dependencies
